// lib/types.ts
export enum Category {
  HEALTH = "HEALTH",
  WORK = "WORK",
  LEARNING = "LEARNING",
  PERSONAL = "PERSONAL",
}

export enum Recurrence {
  DAILY = "DAILY",
  WEEKLY = "WEEKLY",
  MONTHLY = "MONTHLY",
  YEARLY = "YEARLY",
}

export enum Status {
  PENDING = "PENDING",
  DONE = "DONE",
  MISSED = "MISSED",
  SKIPPED = "SKIPPED",
}